Complete impression tracker and implement filter click tracker

+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/reading/data/ProductRatingAndTopic.kt
@@ -41,13 +41,13 @@ data class ProductRating(
         val ratingScore: String = "",
         @SerializedName("totalRating")
         @Expose
-        val totalRating: Int = 0,
+        val totalRating: Long = 0,
         @SerializedName("totalRatingWithImage")
         @Expose
-        val totalRatingWithImage: Int = 0,
+        val totalRatingWithImage: Long = 0,
         @SerializedName("totalRatingTextAndImage")
         @Expose
-        val totalRatingTextAndImage: Int = 0,
+        val totalRatingTextAndImage: Long = 0,
         @SerializedName("detail")
         @Expose
         val detail: List<ProductReviewDetail> = listOf()
